requirements.txt
django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
numpy==1.26.4
